Remove subject preference model
Removes the subject preference model as it is no longer needed. Replaced
by education and teaching_preference models.
