+++ app/models/bookings/placement_request.rb
@@ -3,7 +3,8 @@
 module Bookings
   class PlacementRequest < ApplicationRecord
     include Candidates::Registrations::Behaviours::PlacementPreference
-    include Candidates::Registrations::Behaviours::SubjectPreference
+    include Candidates::Registrations::Behaviours::Education
+    include Candidates::Registrations::Behaviours::TeachingPreference
     include Candidates::Registrations::Behaviours::BackgroundCheck
 
     has_secure_token
