+++ app/services/candidates/registrations/behaviours/education.rb
@@ -1,16 +1,15 @@
 module Candidates
   module Registrations
     module Behaviours
-      module SubjectPreference
+      module Education
         extend ActiveSupport::Concern
 
-        OPTIONS_CONFIG = YAML.load_file "#{Rails.root}/config/candidate_form_options.yml"
-        NOT_APPLYING_FOR_DEGREE = "I don't have a degree and am not studying for one".freeze
-        DEGREE_STAGE_REQUIRING_EXPLANATION = 'Other'.freeze
         NO_DEGREE_SUBJECT = 'Not applicable'.freeze
+        OPTIONS_CONFIG = YAML.load_file "#{Rails.root}/config/candidate_form_options.yml"
+        DEGREE_STAGE_REQUIRING_EXPLANATION = 'Other'.freeze
+        NOT_APPLYING_FOR_DEGREE = "I don't have a degree and am not studying for one".freeze
 
         included do
-          validates :urn, presence: true
           validates :degree_stage, presence: true
           validates :degree_stage, inclusion: { in: :available_degree_stages }, if: -> { degree_stage.present? }
           validates :degree_stage_explaination, presence: true, if: :degree_stage_explaination_required?
@@ -18,21 +17,6 @@ module Candidates
           validates :degree_subject, inclusion: { in: :available_degree_subjects }, if: -> { degree_subject.present? }
           validates :degree_subject, inclusion: [NO_DEGREE_SUBJECT], if: -> { degree_stage.present? && !degree_stage_requires_subject? }
           validates :degree_subject, exclusion: [NO_DEGREE_SUBJECT], if: -> { degree_stage.present? && degree_stage_requires_subject? }
-          validates :teaching_stage, presence: true
-          validates :teaching_stage, inclusion: { in: :available_teaching_stages }, if: -> { teaching_stage.present? }
-          validates :subject_first_choice, presence: true
-          validates :subject_first_choice, inclusion: { in: :available_subject_choices }, if: -> { subject_first_choice.present? }
-          validates :subject_second_choice, presence: true
-          validates :subject_second_choice, inclusion: { in: :second_subject_choices }, if: -> { subject_second_choice.present? }
-        end
-
-        def available_subject_choices
-          @available_subject_choices ||= get_available_subject_choices
-        end
-
-        def second_subject_choices
-          ["I don't have a second subject"] +
-            available_subject_choices
         end
 
         def available_degree_stages
@@ -43,16 +27,12 @@ module Candidates
           OPTIONS_CONFIG.fetch 'DEGREE_SUBJECTS'
         end
 
-        def available_teaching_stages
-          OPTIONS_CONFIG.fetch 'TEACHING_STAGES'
+        def requires_subject_for_degree_stage?(some_degree_stage)
+          some_degree_stage != NOT_APPLYING_FOR_DEGREE
         end
 
         def requires_explanation_for_degree_stage?(some_degree_stage)
           some_degree_stage == DEGREE_STAGE_REQUIRING_EXPLANATION
-        end
-
-        def requires_subject_for_degree_stage?(some_degree_stage)
-          some_degree_stage != NOT_APPLYING_FOR_DEGREE
         end
 
         def no_degree_subject
@@ -67,18 +47,6 @@ module Candidates
 
         def degree_stage_explaination_required?
           requires_explanation_for_degree_stage? degree_stage
-        end
-
-        def get_available_subject_choices
-          if school.subjects.any?
-            school.subjects.pluck(:name)
-          else
-            all_available_subject_choices
-          end
-        end
-
-        def all_available_subject_choices
-          Candidates::School.subjects.map(&:last)
         end
       end
     end
