+++ spec/services/candidates/registrations/education_spec.rb
@@ -1,6 +1,8 @@
 require 'rails_helper'
 
 describe Candidates::Registrations::Education, type: :model do
+  it_behaves_like 'a registration step'
+
   context 'attributes' do
     it { is_expected.to respond_to :degree_stage }
     it { is_expected.to respond_to :degree_stage_explaination }
